Moved all actions out of analysis as the purpose of this module was not very clear
