--- src/analysis/module.type
@@ -1,1 +1,0 @@
-default-on
