+++ src/wham/Makefile
@@ -1,0 +1,4 @@
+USE=colvar core tools matrixtools gridtools generic valtools function
+
+#generic makefile
+include ../maketools/make.module
