+++ src/wham/WhamHistogram.cpp
@@ -25,7 +25,7 @@
 #include "core/ActionRegister.h"
 
 namespace PLMD {
-namespace analysis {
+namespace wham {
 
 //+PLUMEDOC REWEIGHTING WHAM_HISTOGRAM
 /*
